fix timer handling and singleton servie stopping
The timer now actually prints the current time, instead of the time
when it was initialized. Also, when the singleton service is being
stopped, it also stops the timer properly (this wasn't done before
because if the server is going down, the timers will cease to exist
anyway, but stopping them is the right thing to do).

Also, an IllegalStateException from TestingSingletonService.stop
has been prevented -- the method used to call getValue() after
switching "started" from "true" to "false", but getValue() throws
IllegalStateException when "started" is "false".

And finally, JNDI lookups and logging have been cleaned up a bit.